Gradient-Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
duckdb==0.9.2
google-api-python-client==2.193.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
openai==1.30.5
duckduckgo-search==5.3.0
ddgs==9.13.1
httpx==0.27.2
